Reorganize Seo component
