+++ src/pages/credits.js
@@ -2,7 +2,7 @@ import Credits from 'components/credits'
 import Layout from 'components/Layout'
 import PageHeader from 'components/page-header'
 import React from 'react'
-import SEO from 'components/seo'
+import SEO from 'components/Seo'
 import {
   graphql,
   useStaticQuery,
