+++ src/pages/index.js
@@ -3,7 +3,7 @@ import Layout from 'components/Layout'
 import PageHeader from 'components/page-header'
 import ParagraphStyles from 'styles/ParagraphStyles'
 import React from 'react'
-import SEO from 'components/seo'
+import SEO from 'components/Seo'
 import { SectionStyles } from 'styles'
 import Social from 'components/social'
 import styled from 'styled-components'
