+++ src/pages/projects.js
@@ -5,7 +5,7 @@ import ParagraphStyles from 'styles/ParagraphStyles'
 import Projects from 'components/projects'
 import PropTypes from 'prop-types'
 import React from 'react'
-import SEO from 'components/seo'
+import SEO from 'components/Seo'
 import Social from 'components/social'
 import { graphql } from 'gatsby'
