+++ src/pages/uses.js
@@ -3,7 +3,7 @@ import { Link } from 'gatsby'
 import PageHeader from 'components/page-header'
 import ParagraphStyles from 'styles/ParagraphStyles'
 import React from 'react'
-import SEO from 'components/seo'
+import SEO from 'components/Seo'
 import SmallList from 'styles/SmallList'
 import Social from 'components/social'
 import styled from 'styled-components'
